+++ backend/canisters/neuron_controller/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Periodically refresh neuron voting power ([#7604](https://github.com/open-chat-labs/open-chat/pull/7604))
+
 ### Changed
 
 - Log total instructions consumed at end of upgrade ([#7551](https://github.com/open-chat-labs/open-chat/pull/7551))